# Optimise concurrent block production