## Issue Addressed

Closes #5365

## Proposed Changes

Speed up concurrent block production by being stricter about when locks are dropped. We were getting screwed by Rust's auto-drop logic which was not releasing locks in a timely way when they were obtained in the branches of an `if`.

Prior to making this change you can see how staggered the block responses end up (running `blockdreamer` with 4x copies of the same node):

```
Mar 07 06:08:00.111 DEBG Processed HTTP API request, method: GET, path: /eth/v3/validator/blocks/7773640, status: 200 OK, elapsed: 111.176792ms, module: http_api:205
Mar 07 06:08:00.414 DEBG Processed HTTP API request, method: GET, path: /eth/v3/validator/blocks/7773640, status: 200 OK, elapsed: 413.336008ms, module: http_api:205
Mar 07 06:08:00.715 DEBG Processed HTTP API request, method: GET, path: /eth/v3/validator/blocks/7773640, status: 200 OK, elapsed: 714.450081ms, module: http_api:205
Mar 07 06:08:01.027 DEBG Processed HTTP API request, method: GET, path: /eth/v3/validator/blocks/7773640, status: 200 OK, elapsed: 1.026413336s, module: http_api:205
```

Whereas after making this change (mainly the `block_production_state` locking change), they execute properly in parallel:

```
Mar 07 06:23:48.126 DEBG Processed HTTP API request, method: GET, path: /eth/v3/validator/blocks/7773719, status: 200 OK, elapsed: 124.879958ms, module: http_api:205
Mar 07 06:23:48.437 DEBG Processed HTTP API request, method: GET, path: /eth/v3/validator/blocks/7773719, status: 200 OK, elapsed: 436.259608ms, module: http_api:205
Mar 07 06:23:48.449 DEBG Processed HTTP API request, method: GET, path: /eth/v3/validator/blocks/7773719, status: 200 OK, elapsed: 447.748269ms, module: http_api:205
Mar 07 06:23:48.461 DEBG Processed HTTP API request, method: GET, path: /eth/v3/validator/blocks/7773719, status: 200 OK, elapsed: 459.995933ms, module: http_api:205
```

The first request is 300ms faster than the subsequent 3 because it is able to use the single-use `block_production_state` cache.