@michaelsproul 2024-03-07
Not fully satisfied with the auto-drop explanation. In isolation Rust seems to be smarter than this: https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=5418647540cf01a97beda20484399464