@michaelsproul 2024-03-08
I found out some interesting things about the drop ordering. If the lock guard is borrowed _mutably_ then the compiler will drop it "late". In my first example above we don't see the late drop behaviour because the borrow is immutable.

Copying the structure of Lighthouse's code more closely reproduces the issue:

https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=3bcdbd11c7aa3e8ea5e9ce750e12e723

This prints:

```
case 2: World
releasing the lock lock2
releasing the lock lock1
```

Modifying the first playground to do a mutable borrow _only_ for the first lock, forces that lock to be dropped late, while the other lock is dropped early:

https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=a7689777c1941e424d91b1d1e4728c12

```
releasing the lock lock2
case 2: World
releasing the lock lock1
```

I think this could be a good Clippy lint!